Supporting community with data
Inspired by data suggesting the importance of community as it correlates to success and health, this collaboration of four data scientists identified markers of residential areas with reciprocal support between citizens and the community itself. We utilized both national and local data, which in our case is San Diego, California.